+++ scripts/macros/2024/classFeatures/cleric/war/avatarOfBattle.js
@@ -1,0 +1,5 @@
+export let avatarOfBattle = {
+    name: 'Avatar of Battle',
+    version: '1.3.31',
+    rules: 'modern'
+};